Used single quotes and merged the ChatBox and index file
