--- src/pages/components/chatbot/ChatBox.js
@@ -1,33 +1,0 @@
-import React, { Fragment } from 'react';
-import Chatbot from 'react-chatbot-kit';
-import 'react-chatbot-kit/build/main.css';
-
-import '../../../assets/css/chatbot.css';
-
-import chatbotConfig from './configs/chatbotConfig';
-import ActionProvider from './configs/ActionProvider';
-import MessageParser from './configs/MessageParser';
-
-const ChatBox = ({ isChatBoxOpen, closeChatBox }) => {
-  const loadMessages = () => {
-    const messages = JSON.parse(localStorage.getItem('chat_messages'));
-
-    return messages;
-  };
-
-  return (
-    <Fragment>
-      {isChatBoxOpen && (
-        <Chatbot
-          config={chatbotConfig(closeChatBox)}
-          actionProvider={ActionProvider}
-          messageHistory={loadMessages()}
-          messageParser={MessageParser}
-          placeholderText="Send a message..."
-        />
-      )}
-    </Fragment>
-  );
-};
-
-export default ChatBox;
