+++ src/pages/components/chatbot/index.js
@@ -1,5 +1,12 @@
 import React, { Fragment, useState } from 'react';
-import ChatBox from './ChatBox';
+import Chatbot from 'react-chatbot-kit';
+import 'react-chatbot-kit/build/main.css';
+
+import '../../../assets/css/chatbot.css';
+
+import chatbotConfig from './configs/chatbotConfig';
+import ActionProvider from './configs/ActionProvider';
+import MessageParser from './configs/MessageParser';
 
 const ContactUs = () => {
   const [isChatBoxOpen, setIsChatBoxOpen] = useState(false);
@@ -8,21 +15,35 @@ const ContactUs = () => {
     localStorage.removeItem('chat_messages');
   });
 
+  const closeChatBox = () => setIsChatBoxOpen(false);
+
+  const loadMessages = () => {
+    const messages = JSON.parse(localStorage.getItem('chat_messages'));
+
+    return messages;
+  };
+
   return (
     <Fragment>
-      <ChatBox
-        isChatBoxOpen={isChatBoxOpen}
-        closeChatBox={() => setIsChatBoxOpen(false)}
-      />
+      {isChatBoxOpen && (
+        <Chatbot
+          config={chatbotConfig(closeChatBox)}
+          actionProvider={ActionProvider}
+          messageHistory={loadMessages()}
+          messageParser={MessageParser}
+          placeholderText='Send a message...'
+        />
+      )}
+
       <img
-        className="z-50 cursor-pointer fixed bottom-0 right-0 mb-4 mr-6 w-[50px] md:mb-7 md:mr-[2rem] lg:w-[4.5rem] lg:mb-7"
+        className='z-50 cursor-pointer fixed bottom-0 right-0 mb-4 mr-6 w-[50px] md:mb-7 md:mr-[2rem] lg:w-[4.5rem] lg:mb-7'
         onClick={() => setIsChatBoxOpen(!isChatBoxOpen)}
         src={
           isChatBoxOpen
             ? '/images/CloseFloatingButton.svg'
             : '/images/ChatFloatingButton.svg'
         }
-        alt="chat floating button"
+        alt='chat floating button'
       />
     </Fragment>
   );
